completed the assignment

+++ problem1.cpp
@@ -17,7 +17,7 @@ void extractor(int n, int* arr) {
     }
 }
 
-int smallestNumber(int *arr) {
+long long smallestNumber(int *arr) {
     int small=0;
     for(int i=0;i<10;i++) {
         if(arr[i]>0) {
